% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_object.R
\name{as.data.frame.xrd}
\alias{as.data.frame.xrd}
\title{Convert xrd data to a data frame}
\usage{
\method{as.data.frame}{xrd}(x, ...)
}
\arguments{
\item{x}{xrd S3 class}

\item{...}{dots}
}
\description{
Convert xrd data to a data frame
}
