% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_object.R
\name{create_xrd}
\alias{create_xrd}
\title{Create an xrd S3 object}
\usage{
create_xrd(TwoTheta, I, I.meas, loop)
}
\arguments{
\item{TwoTheta}{vector of two theta angles from xrd spectrum}

\item{I}{intensity (counts per second)}

\item{I.meas}{measured intentisty}

\item{loop}{integer of a specific loop}
}
\description{
Create an xrd S3 object
}
