% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2genx.R
\name{export2genx}
\alias{export2genx}
\title{Export XRD data to GenX format}
\usage{
export2genx(dataXRD, exportPath = ".", minTheta = 0, maxTheta = 4)
}
\arguments{
\item{dataXRD}{xrd data from a data package}

\item{minTheta}{minimum 2q to be included}

\item{maxTheta}{maximum 2q to be included in export}
}
\value{
file name for export
}
\description{
Export XRD data to GenX format
}
\examples{
export2genx(7)

}
