% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.import.R
\name{plot.xrd}
\alias{plot.xrd}
\title{Plots xrd class}
\usage{
\method{plot}{xrd}(x, ...)
}
\arguments{
\item{x}{xrd S3 object}

\item{...}{Optional graphical parameters to adjust different components of the performance plot}
}
\description{
Plots xrd class
}
