% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_object.R
\name{print.xrd}
\alias{print.xrd}
\title{Print xrd object as table}
\usage{
\method{print}{xrd}(x, ...)
}
\arguments{
\item{x}{xrd object.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print xrd object as table
}
