% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.find.Peak.R
\name{xrd.find.Peak}
\alias{xrd.find.Peak}
\title{Finds a peak}
\usage{
xrd.find.Peak(TwoTheta, Intensity, PeakPos)
}
\arguments{
\item{TwoTheta}{angle}

\item{Intensity}{intensity signal}

\item{PeakPos}{approximate angle of peak position}
}
\value{
fit peak position
}
\description{
Finds a peak
}
\examples{
filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
d = xrd.read.ASC(filename)
xrd.find.Peak(d$theta, d$I, 38.2)

}
