% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.FringeAnalysis.R
\name{xrd.FringeAnalysis}
\alias{xrd.FringeAnalysis}
\title{XRD / XRR fringe analysis}
\usage{
xrd.FringeAnalysis(dataXRD)
}
\arguments{
\item{dataXRD}{data frame with theta (2Theta) and I (intensity)}
}
\value{
list with thickness (`t.nm`), table of peaks (`df.peak`), and two ggplot graphs (`g.peaks` and `g.thickness`)
}
\description{
Searches for fringes in the XRD or XRR spectrum and
returns a list with the thickness, a peak table, and
two graphs displaying the results.
}
\examples{
filename <- xrd.getSampleFiles('ras')
data <- xrd.import(filename, xrd=TRUE)
data_5 = xrd_filter(data, 5,10)
analysis <- xrd.FringeAnalysis(data_5)
analysis$t.nm

}
\author{
Thomas Gredig
}
