% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.AllPeaks.R
\name{xrd.get.AllPeaks}
\alias{xrd.get.AllPeaks}
\title{Finds all peaks in an XRD spectrum}
\usage{
xrd.get.AllPeaks(
  data,
  min.Prominence = 5,
  Try.Sigma = c(0.2, 0.1, 0.05, 0.3),
  deltaTheta = 5,
  Range = c(0, 90)
)
}
\arguments{
\item{data}{data frame with intensity and 2 theta angle.}

\item{min.Prominence}{minimum prominence in percent (optional).}

\item{Try.Sigma}{vector with peak widths used to start fitting (optional).}

\item{deltaTheta}{search area around main peak (optional).}

\item{Range}{range to search for peaks (optional).}
}
\value{
sorted vector with peak positions with 0.01 degree precision.
}
\description{
Attempts to find peaks in an XRD spectrum and return a list of angles with
the peak locations. (Use with caution, might find additional peaks.)
}
\examples{

filename = xrd.getSampleFiles('asc')
d = xrd.import(filename)
xrd.get.AllPeaks(d, deltaTheta = 2, Try.Sigma = c(0.1),  Range = c(42, 46))

}
\seealso{
[xrd.find.Peak()]
}
