% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.PeakStats.R
\name{xrd.get.PeakStats}
\alias{xrd.get.PeakStats}
\title{Fits the peak to a Gaussian, usually run after a peak is
identified through its location}
\usage{
xrd.get.PeakStats(TwoTheta, Intensity, PeakPos)
}
\arguments{
\item{TwoTheta}{angle}

\item{Intensity}{intensity signal}

\item{PeakPos}{approximate angle of peak position}
}
\value{
background, amplitude, position, width + 4 std. errors
}
\description{
Fits the peak to a Gaussian, usually run after a peak is
identified through its location
}
\examples{
filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
d = xrd.read.ASC(filename)
xrd.get.Peak.Stats(d$theta, d$I, 38.2)

}
