% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.getSampleFiles.R
\name{xrd.getSampleFiles}
\alias{xrd.getSampleFiles}
\title{XRD Filenames with Sample Data}
\usage{
xrd.getSampleFiles(fileExt = NULL)
}
\arguments{
\item{fileExt}{XRD filename extension: "asc", "txt", "ras", "rasx"}
}
\value{
vector with path and filename to sample XRD data files
}
\description{
Returns filenames with XRD sample data stored in different formats.
These data sets help the user quickly get started with example codes
for this package. It can also be used for testing and debugging purposes.
d
}
\examples{
file.list = xrd.getSampleFiles()
basename(file.list)

}
\author{
Thomas Gredig
}
