% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.getSampleFiles.R
\name{xrd.getSamplePath}
\alias{xrd.getSamplePath}
\title{Return sample path for XRD data files}
\usage{
xrd.getSamplePath()
}
\value{
path for sample XRD data files
}
\description{
returns sample XRD path from library
}
\examples{
pfad = xrd.getSamplePath()
}
\author{
Thomas Gredig
}
