% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.importData.R
\name{xrd.importData}
\alias{xrd.importData}
\title{Imports Rigaku XRD data file}
\usage{
xrd.importData(filename)
}
\arguments{
\item{filename}{full filename with path}
}
\value{
data frame with XRD data
}
\description{
Imports Rigaku XRD data file
}
\examples{

filename = xrd.getSampleFiles()[1]
d = xrd.importData(filename)
plot(d$theta,d$I,log='y',col='red')

}
