% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.peakEstimate.R
\name{xrd.peakEstimate}
\alias{xrd.peakEstimate}
\title{Estimate highest peak in selected spectrum}
\usage{
xrd.peakEstimate(TwoTheta, Intensity = NULL)
}
\arguments{
\item{TwoTheta}{vector of two theta angles from xrd spectrum}

\item{Intensity}{\code{NULL} if xrd S3 object, otherwise xrd intensity vector}
}
\value{
4 estimated parameters: background (b0), amplitude (A0), peak position (th0), and half-width (s0)
}
\description{
Simple peak position estimate, yields largest peak
in a data set. This is not a fit, just an
estimate based on the data at hand.
It returns the background (b0), the amplitude (A0), the
peak position (th0), and the half-width sigma (s0).
}
\examples{
d = xrd.import(xrd.getSampleFiles('asc'), TRUE)
d = xrd_filter(d, 30,42)
p = xrd.peakEstimate(d)
plot(d); abline(v=p[3], lwd=2)
abline(h=p[1], col='darkgrey')
p

}
