% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.importData.R
\name{xrd.read.RAS}
\alias{xrd.read.RAS}
\title{Reads the RAS Rigaku XRD data}
\usage{
xrd.read.RAS(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with XRD data
}
\description{
Reads the RAS Rigaku XRD data
}
\examples{

library(dplyr)
filename = dir(xrd.getSamplePath(),patter='ras$')
d = xrd.read.RAS(file.path(xrd.getSamplePath(), filename))
plot(d$theta,d$I,log='y',col='red')

}
