% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.importData.R
\name{xrd.read.RASX}
\alias{xrd.read.RASX}
\title{Reads the RASX Rigaku XRD data}
\usage{
xrd.read.RASX(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with XRD data
}
\description{
Reads the RASX Rigaku XRD data
}
\examples{
fname = xrd.getSampleFiles('rasx')[1]
d = xrd.read.RASX(fname)
head(d)

}
\author{
Thomas Gredig
}
