% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.importData.R
\name{xrd.read.TXT}
\alias{xrd.read.TXT}
\title{Reads the TXT Rigaku XRD file with a Header}
\usage{
xrd.read.TXT(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with XRD data
}
\description{
Reads the TXT Rigaku XRD file with a Header
}
\examples{

fname = xrd.getSampleFiles('txt')
d = xrd.read.TXT(fname)
plot(d$theta,d$I,log='y',col='red')

}
