% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.reader.R
\name{xrd.read.TxtWithHeader}
\alias{xrd.read.TxtWithHeader}
\title{Reads the Text converted XRD file with a Header}
\usage{
xrd.read.TxtWithHeader(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
list with data and header
}
\description{
Reads the Text converted XRD file with a Header
}
\examples{
d = xrd.read.TxtWithHeader(filename)

}
