% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.import.R
\name{xrd.readHeader.ASC}
\alias{xrd.readHeader.ASC}
\title{Reads the header of an XRD ASC file}
\usage{
xrd.readHeader.ASC(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with XRD header
}
\description{
Reads the header of an XRD ASC file
}
