% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd_object.R
\name{xrd_filter}
\alias{xrd_filter}
\title{Filters XRD data}
\usage{
xrd_filter(x, th_min, th_max)
}
\arguments{
\item{x}{xrd S3 class}

\item{th_min}{minimum theta angle}

\item{th_max}{maximum theta angle}
}
\description{
Filters XRD data
}
