% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export2genx.R
\name{export2genx}
\alias{export2genx}
\title{Export XRD Data into GenX Format}
\usage{
export2genx(
  dataXRD,
  exportPath = NULL,
  minTheta = 0,
  maxTheta = 4,
  GenX_filename = "xrd_data_genx_format.txt"
)
}
\arguments{
\item{dataXRD}{xrd Object}

\item{exportPath}{default is a temporary direction, stores the output file in this directory}

\item{minTheta}{minimum 2Theta to be included}

\item{maxTheta}{maximum 2Theta to be included in export}

\item{GenX_filename}{filename for output file}
}
\value{
file name of the exported GenX text file
}
\description{
GenX uses the differential evolution algorithm for fitting X-ray
reflectivity data. This function normalizes the data and formats
x-ray reflectivity data, so that it can be imported with the GenX
code. The data is saved to a text file. More details are found at:
https://doi.org/10.1107/S1600576722006653
}
\examples{
filename = xrd.getSampleFiles(fileExt='asc')
d = xrd.import(filename, dataXRD=TRUE)
peak.pos = xrd.find.Peak(d, peakPos = 38.2)
plot(d)
abline(v=peak.pos,col='blue')

}
\seealso{
[xrd.import()]
}
