% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.find.Peak.R
\name{xrd.find.Peak}
\alias{xrd.find.Peak}
\title{Finds peak position from gaussian fit}
\usage{
xrd.find.Peak(
  TwoTheta,
  Intensity = NULL,
  peakPos = NA,
  Try.Sigma = c(0.1, 0.4, 0.2, 0.15),
  thetaDelta = 5
)
}
\arguments{
\item{TwoTheta}{vector of two theta angles from xrd spectrum.}

\item{Intensity}{\code{NULL} if xrd S3 object, otherwise xrd intensity vector.}

\item{peakPos}{approximate angle (2Theta) of peak position.}

\item{Try.Sigma}{vector with peak widths used to start fitting.}

\item{thetaDelta}{approximate width of angle to search peak.}
}
\value{
peak position angle obtained from fit
}
\description{
Fits a Gaussian form to a peak
}
\examples{
filename = xrd.getSampleFiles(fileExt='asc')
d = xrd.import(filename)
peak.pos = xrd.find.Peak(d$TwoTheta, d$I, 38.2)
plot(d$TwoTheta,d$I,log='y',col='red')
abline(v=peak.pos,col='blue')

}
