% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.find.Peak.R
\name{xrd.find.Peak}
\alias{xrd.find.Peak}
\title{Finds Peak Position from Gaussian}
\usage{
xrd.find.Peak(
  TwoTheta,
  Intensity,
  peakPos = NA,
  Try.Sigma = c(0.1, 0.4, 0.2, 0.15),
  thetaDelta = 5,
  verbose = FALSE
)
}
\arguments{
\item{TwoTheta}{angle}

\item{Intensity}{intensity signal}

\item{peakPos}{approximate angle of peak position}

\item{Try.Sigma}{vector with peak widths used to start fitting}

\item{thetaDelta}{width of angle to search peak}

\item{verbose}{logical, if \code{TRUE} provides extra information}
}
\value{
fit peak position
}
\description{
Fits a Gaussian form to a peak
}
\examples{
filename = xrd.getSampleFiles(fileExt='asc')
d = xrd.import(filename)
peak.pos = xrd.find.Peak(d$theta, d$I, 38.2)
plot(d$theta,d$I,log='y',col='red')
abline(v=peak.pos,col='blue')

}
