% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.find.Peak.R
\name{xrd.find.Peak}
\alias{xrd.find.Peak}
\title{Finds Peak Position from Gaussian}
\usage{
xrd.find.Peak(TwoTheta, Intensity, PeakPos, Try.Sigma = c(0.1, 0.2, 0.15))
}
\arguments{
\item{TwoTheta}{angle}

\item{Intensity}{intensity signal}

\item{PeakPos}{approximate angle of peak position}

\item{Try.Sigma}{vector with peak widths used to start fitting}
}
\value{
fit peak position
}
\description{
Fits a Gaussian form to a peak
}
\examples{
library(stringr)
filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
d = xrd.read.ASC(filename)
plot(d$theta,d$I,log='y',col='red')
peak.pos = xrd.find.Peak(d$theta, d$I, 38.2)
abline(v=peak.pos,col='blue')

}
