% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.FringeAnalysis.R
\name{xrd.FringeAnalysis}
\alias{xrd.FringeAnalysis}
\title{XRD / XRR Fringe Analysis}
\usage{
xrd.FringeAnalysis(df)
}
\arguments{
\item{df}{data frame with theta (2q) and I (intensity)}
}
\description{
XRD / XRR Fringe Analysis
}
