% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.FringeAnalysis.R
\name{xrd.FringeAnalysis}
\alias{xrd.FringeAnalysis}
\title{XRD / XRR Fringe Analysis}
\usage{
xrd.FringeAnalysis(dataXRD)
}
\arguments{
\item{dataXRD}{data frame with theta (2Theta) and I (intensity)}
}
\value{
list with thickness, table of peaks, and two ggplot graphs
}
\description{
Searches for fringes in the XRD or XRR spectrum and
returns a list with the thickness, a peak table, and
two graphs displaying the results.
}
\examples{
filename <- xrd.getSampleFiles('asc')
data <- xrd.import(filename, dataXRD=TRUE)
data_35 = xrd_filter(data, 35,50)
plot(data_35)
analysis <- xrd.FringeAnalysis(data_35)

}
\author{
Thomas Gredig
}
