% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.FringeThickness.R
\name{xrd.FringeThickness}
\alias{xrd.FringeThickness}
\title{Fringe Thickness Calculation}
\usage{
xrd.FringeThickness(theta1, theta2, lambda = 1.5406)
}
\arguments{
\item{theta1}{first Two Theta angle in degree}

\item{theta2}{second Two Theta angle in degree}

\item{lambda}{wave length of x-ray radiation in units of Angstrom}
}
\value{
thickness in units of Angstrom
}
\description{
Fringe Thickness Calculation
}
\examples{
print("Should be 111.64 nm")
xrd.FringeThickness(66.22698, 66.32140)

}
