% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.AllPeaks.R
\name{xrd.get.AllPeaks}
\alias{xrd.get.AllPeaks}
\title{Finds all the peaks in a spectrum}
\usage{
xrd.get.AllPeaks(TwoTheta, Intensity, Try.Sigma = c(0.1, 0.2, 0.15),
  Range = c(0, 90))
}
\arguments{
\item{TwoTheta}{angle}

\item{Intensity}{intensity signal}

\item{Try.Sigma}{vector with peak widths used to start fitting (optional)}

\item{Range}{range to search for peaks (optional)}
}
\value{
vector with peak positions
}
\description{
Finds all the peaks in a spectrum
}
\examples{

filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
d = xrd.read.ASC(filename)
peak.list = xrd.get.AllPeaks(d$theta, d$I)

}
