% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.AllPeaks.R
\name{xrd.get.AllPeaks}
\alias{xrd.get.AllPeaks}
\title{Finds All Peaks in a Spectrum}
\usage{
xrd.get.AllPeaks(
  data,
  min.Prominence = 5,
  Try.Sigma = c(0.2, 0.1, 0.05, 0.3),
  deltaTheta = 5,
  Range = c(0, 90),
  verbose = FALSE
)
}
\arguments{
\item{data}{data frame with intensity and 2 theta angle}

\item{min.Prominence}{minimum prominence in percent (optional)}

\item{Try.Sigma}{vector with peak widths used to start fitting (optional)}

\item{deltaTheta}{search area around main peak}

\item{Range}{range to search for peaks (optional)}

\item{verbose}{logical, if \code{TRUE} outputs additional info}
}
\value{
vector with peak positions
}
\description{
Attempts to find peaks in an XRD spectrum and return a list of angles with
the peak locations. (This function is highly experimental and does not
work well.)
}
\examples{

filename = xrd.getSampleFiles('asc')
d = xrd.read.ASC(filename)
xrd.get.AllPeaks(d, deltaTheta = 2, Try.Sigma = c(0.1),  Range = c(42, 46))

}
\seealso{
[xrd.find.Peak()]
}
