% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.DebyeScherrer.R
\name{xrd.get.DebyeScherrer}
\alias{xrd.get.DebyeScherrer}
\title{Returns size estimate from Debye-Schrerrer Equation}
\usage{
xrd.get.DebyeScherrer(PeakStats, Lambda = 1.5406)
}
\arguments{
\item{PeakStats}{vector with the peak statistics}

\item{Lambda}{wavelength of x-ray in (A)}
}
\value{
peak width in Angstrom
}
\description{
Returns size estimate from Debye-Schrerrer Equation
}
\examples{
filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
d = xrd.read.ASC(filename)
xrd.get.DebyeScherrer(d$theta, d$I, 38.2)

}
