% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.PeakProminence.R
\name{xrd.get.PeakProminence}
\alias{xrd.get.PeakProminence}
\title{Returns the Peak Prominence}
\usage{
xrd.get.PeakProminence(PeakStats)
}
\arguments{
\item{PeakStats}{vector with the peak statistics}
}
\value{
percentage of the peak height with respect to background
}
\description{
Returns the Peak Prominence
}
\examples{
filename = xrd.getSampleFiles()[1]
d = xrd.read.ASC(filename)
q = xrd.get.PeakStats(d$theta, d$I, 38.2)
xrd.get.PeakProminence(q)

}
