% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.PeakStats.R
\name{xrd.get.PeakStats}
\alias{xrd.get.PeakStats}
\title{Fits a Gaussian function to XRD data}
\usage{
xrd.get.PeakStats(
  TwoTheta,
  Intensity = NULL,
  peakPos,
  Try.Sigma = c(0.2, 0.1, 0.05, 0.3)
)
}
\arguments{
\item{TwoTheta}{vector of two theta angles from xrd spectrum.}

\item{Intensity}{\code{NULL} if xrd S3 object, otherwise xrd intensity vector.}

\item{peakPos}{approximate 2Theta angle of peak position.}

\item{Try.Sigma}{vector with peak widths used to start fitting.}
}
\value{
background, amplitude, position, width + 4 std. errors from peak fit
}
\description{
Given an XRD peak, a Gaussian fit at the provided location \code{peakPos}
a Gaussian fit is made to find the exact location, width, and background.
}
\examples{
filename = xrd.getSampleFiles(fileExt='asc')
d = xrd.import(filename)
xrd.get.PeakStats(d$TwoTheta, d$I, 38.2)

}
\author{
Thomas Gredig
}
