% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.PeakStats.R
\name{xrd.get.PeakStats}
\alias{xrd.get.PeakStats}
\title{Fits the peak to a Gaussian}
\usage{
xrd.get.PeakStats(TwoTheta, Intensity, PeakPos, Try.Sigma = c(0.1, 0.2, 0.15))
}
\arguments{
\item{TwoTheta}{angle}

\item{Intensity}{intensity signal}

\item{PeakPos}{approximate angle of peak position}

\item{Try.Sigma}{vector with peak widths used to start fitting}
}
\value{
background, amplitude, position, width + 4 std. errors
}
\description{
usually run after a peak is
identified through its location
}
\examples{
filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
d = xrd.read.ASC(filename)
xrd.get.PeakStats(d$theta, d$I, 38.2)

}
