% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.get.PeakStats.R
\name{xrd.get.PeakStats}
\alias{xrd.get.PeakStats}
\title{Fits a Gaussian Function to XRD data}
\usage{
xrd.get.PeakStats(
  TwoTheta,
  Intensity,
  peakPos,
  Try.Sigma = c(0.2, 0.1, 0.05, 0.3),
  verbose = FALSE
)
}
\arguments{
\item{TwoTheta}{2q angle}

\item{Intensity}{XRD intensity}

\item{peakPos}{specific 2q angle of peak position}

\item{Try.Sigma}{vector with peak widths used to start fitting}

\item{verbose}{logical, if \code{TRUE} output additional information}
}
\value{
background, amplitude, position, width + 4 std. errors
}
\description{
usually run after a peak is
identified through its location
}
\examples{
filename = xrd.getSampleFiles(fileExt='asc')
d = xrd.read.ASC(filename)
xrd.get.PeakStats(d$theta, d$I, 38.2)

}
\author{
Thomas Gredig
}
