% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.getSampleFiles.R
\name{xrd.getSampleFiles}
\alias{xrd.getSampleFiles}
\title{Returns XRD sample files}
\usage{
xrd.getSampleFiles(fileExt = NA)
}
\arguments{
\item{fileExt}{file extension, could be \code{NA} for all, or "asc", "txt", "ras", "rasx"}
}
\value{
vector with path/filename to sample XRD data files
}
\description{
returns sample XRD  from library
}
\examples{
file.list = xrd.getSampleFiles()
print(paste("Found",length(file.list),"sample files:"))
basename(file.list)

}
\author{
Thomas Gredig
}
