% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.importData.R
\name{xrd.import}
\alias{xrd.import}
\title{Imports Rigaku XRD Data File}
\usage{
xrd.import(filename, dataXRD = FALSE)
}
\arguments{
\item{filename}{full file name with path}

\item{dataXRD}{logical, if \code{TRUE} new xrd S3 class is returned}
}
\value{
data frame with XRD data
}
\description{
Import function recognizes ASC, TXT, RAS, and RASX files from Rigaku XRD
instrument; returns dataframe with 2theta, I (intensity normalized for time),
and I.meas (measured intensity)
}
\examples{
fname = xrd.getSampleFiles('rasx')
d = xrd.import(fname, TRUE)
plot(d)

}
\author{
Thomas Gredig
}
