% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.importData.R
\name{xrd.import}
\alias{xrd.import}
\title{Imports Rigaku XRD data file}
\usage{
xrd.import(filename)
}
\arguments{
\item{filename}{full filename with path}
}
\value{
data frame with XRD data
}
\description{
Import function recognizes ASC, TXT, RAS, and RASX files from Rigaku XRD
instrument; returns dataframe with 2theta, I (intensity normalized for time),
and I.meas (measured intentsity)
}
\examples{

fname = xrd.getSampleFiles()[1]
d = xrd.import(fname)
plot(d$theta,d$I,log='y',col='red')

}
\author{
Thomas Gredig
}
