% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.reader.R
\name{xrd.read.ASC}
\alias{xrd.read.ASC}
\title{Reads the ASC Rigaku XRD file with a Header}
\usage{
xrd.read.ASC(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with XRD data
}
\description{
Reads the ASC Rigaku XRD file with a Header
}
\examples{

filename = system.file("extdata", "2Theta.asc", package='rigakuXRD')
d = xrd.read.ASC(filename)

}
