% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.read.RAS.R
\name{xrd.read.RAS}
\alias{xrd.read.RAS}
\title{Reads the RAS Rigaku XRD data}
\usage{
xrd.read.RAS(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with XRD data
}
\description{
Reads the RAS Rigaku XRD data
}
\examples{

filename = system.file("extdata", "2Theta.ras", package='rigakuXRD')
d = xrd.read.RAS(filename)

}
