% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrd.importData.R
\name{xrd.read.TXTnoheader}
\alias{xrd.read.TXTnoheader}
\title{Reads the TXT Rigaku XRD file with no header}
\usage{
xrd.read.TXTnoheader(filename)
}
\arguments{
\item{filename}{filename including path}
}
\value{
data frame with XRD data and columns TwoTheta and I
}
\description{
Reads the TXT Rigaku XRD file with no header
}
\examples{

filename = dir(xrd.getSamplePath(),patter='txt$')[1]
d = xrd.read.TXTnoheader(file.path(xrd.getSamplePath(), filename))
plot(d$TwoTheta,d$I,log='y',col='red')

}
